requirements.txt
pandas~=1.3.5
python-dateutil~=2.8.2
openpyxl~=3.0.9
pytz~=2021.3
numpy~=1.22.1
setuptools~=57.0.0
PyQt6~=6.1.0
dateparser~=1.1.1